fastapideta/transport-app-api/requirements.txt
fastapi==0.85.1
uvicorn==0.19.0
requests==2.28.1
python-dotenv==0.21.0
pandas==1.5.1